Replace the old Ufo module with the new one
